Conflicts:
	src/weather_gen.cpp

+++ src/weather_gen.cpp
@@ -1,6 +1,4 @@
 #include "weather_gen.h"
-#include "PerlinNoise/PerlinNoise.hpp"
-#include "calendar.h"
 #include "options.h"
 #include <cmath>
 #include <iostream>
@@ -21,11 +19,14 @@ weather_generator::weather_generator(unsigned seed):
 w_point weather_generator::get_weather(double x, double y, calendar t) {
     double z = (double) t.get_turn() / 2000.0; // Integer turn / widening factor of the Perlin function.
     double day_of_year = t.days() +  static_cast<double>(OPTIONS["SEASON_LENGTH"]) * t.get_season();
-    double dayFraction(t.minutes_past_midnight() / 1440);
+    double dayFraction((double)t.minutes_past_midnight() / 1440);
+
+    // Noise factors
     double T(Temperature.noise(x, y, z) * 8.0);
     double H(Humidity.noise(x, y, z / 5));
     double H2(Humidity.noise(x, y, z) / 4);
     double P(Pressure.noise(x, y, z / 3) * 70);
+
     // temperature variation
     double now((day_of_year + (dayFraction)) / this->year_length); // Add the minutes to the day and return the current time as a decimal [0-1].
     double mod_t(0); // TODO: make this depend on latitude and altitude?
@@ -34,13 +35,17 @@ w_point weather_generator::get_weather(double x, double y, calendar t) {
     double season_atenuation(cos( tau * now ) / 2 + 1); // Harsh winter nights, hot summers.
     double season_dispersion(pow(2, cos( tau * now ) * -1 + 1) - 2.3); // Make summers peak faster and winters not perma-frozen.
     double daily_variation(cos( tau * dayFraction ) * season_atenuation + season_dispersion); // Day-night temperature variation.
+
     T += current_t; // Add baseline to the noise.
     T += seasonal_variation * 12 * exp(-pow(current_t * 2.7 / 20 - 0.5, 2)); // Add season curve offset to account for the winter-summer difference in day-night difference.
     T += daily_variation * 10 * exp(-pow(current_t / 30, 2)); //((4000 / (current_t + 115) - 24) + seasonal_variation); // Add daily variation scaled to the inverse of the current baseline. A very specific and finicky adjustment curve.
+
     // humidity variation
+
     double mod_h(0);
     double current_h(this->base_h + mod_h);
     H = std::max(std::min((cos( tau * now ) / 10.0 + (-pow(H, 2)*3 + H2)) * current_h/2.0 + current_h, 100.0), 0.0); // Humidity stays mostly at the mean level, but has low peaks rarely. It's a percentage.
+
     // pressure variation
     double mod_p(0);
     double current_p(this->base_p + mod_p);
